mqtt-data-normalization/requirements.txt
quixstreams==3.21.0
python-dotenv